+++ components/LayoutOperationsNav.vue
@@ -1,23 +1,42 @@
 <template>
-  <aside class="module-nav" :style="{height:`${mainHeight()}px`}">
-    <ul class="module-features text-center">
-      <li v-if="!activeOperations.length">
-        <nuxt-link to="/add">Chose an operation</nuxt-link>
-      </li>
-      <li v-for="operation in activeOperations" :key="operation.id" :class="isActiveViewClass(operation.id)" class="module-feature">
-        <nuxt-link :to="`/edit?id=${operation.id}`" class="feature-link pointer">
-          <span class="feature-icon"><fa :icon="['fas', operation.icon]" /></span>
-          <span>{{ operation.name }}</span>
+  <aside class="operation-nav" :style="{height:`${mainHeight()}px`}">
+    <ul class="operation-features text-center">
+      <li v-for="(operation, index) in activeOperations" :key="operation.id" :class="isActiveViewClass(operation.id)" class="operation-feature">
+        <nuxt-link :to="`/edit?id=${operation.id}`" class="operation-item pointer">
+          <div class="operation-title">
+            <span class="feature-icon pull-right">
+              <fa
+                @click="removeOperation(operation.id)"
+                :icon="['fas', 'ellipsis-vertical']" />
+            </span>
+            <span class="feature-icon"><fa :icon="['fas', categoryToIcon(operation.category)]" /></span>
+            <span>{{ operation.name }}</span>
+          </div>
+          <span class="operation-next">
+            <fa :icon="['fas', 'arrow-down-long']" />
+            <span class="operation-next-add">
+              <nuxt-link class="add-action" :to="`/add?after=${index+1}`">
+                <fa :icon="['fas', 'plus']" class="add-action-icon" />
+              </nuxt-link>
+            </span>
+          </span>
         </nuxt-link>
       </li>
-      <li v-if="activeOperations.length">
-        <nuxt-link to="/add">Next step</nuxt-link>
+    </ul>
+    <ul class="operation-features text-center">
+      <li>
+        <nuxt-link class="add-action" to="/add">
+          <fa :icon="['fas', 'plus']" class="add-action-icon" />
+        </nuxt-link>
       </li>
     </ul>
   </aside>
 </template>
 
 <script>
+import { mapActions } from 'vuex'
+import { Utils } from '@/services/Utils'
+
 export default {
   name: 'LayoutOperationsNav',
   computed: {
@@ -29,6 +48,12 @@ export default {
     }
   },
   methods: {
+    ...mapActions({
+      removeOperation: 'removeOperation'
+    }),
+    categoryToIcon (category) {
+      return Utils.categoryToIcon(category)
+    },
     isActiveViewClass (viewCode) {
       return this.$store.state.activeView === `${this.activeModuleCode}:${viewCode}` ? 'active' : false
     },
@@ -41,7 +66,21 @@ export default {
 }
 </script>
 <style>
-.module-nav{
+.add-action{
+  margin: 10px auto;
+  display: inline-block;
+  font-size: 2em;
+  border-radius: 50%;
+  width: 40px;
+  height: 40px;
+  line-height: 38px;
+  color: #1155cb;
+  border: 1px solid #1155cb;
+}
+add-action-icon{
+  font-size: 2em;
+}
+.operation-nav{
   display: block;
   float: left;
   height: 100%;
@@ -52,57 +91,78 @@ export default {
   background: #ffffff;
   border-right: 1px solid #1155cb;
 }
-.module-nav-title{
-  text-transform: uppercase;
-  margin-top: 10px;
-  margin-bottom: 20px;
-
-}
-.module-nav-inner{
-  padding: 20px;
-}
-.module-features{
+.operation-features{
   list-style: none;
   padding: 0;
   float: left;
   width: 100%;
 }
-.module-feature{
+.operation-feature{
   display: inline-block;
   background: #ffffff;
   float: left;
-  border-bottom: 1px solid #1155cb;
   position: relative;
   width: 100%;
 }
-.module-features > .module-feature > .feature-link {
+.operation-features > .operation-feature > .operation-item {
   text-transform: uppercase;
   font-weight: 700;
 }
-.module-subfeature{
-  display: inline-block;
-  float: left;
-  position: relative;
-  width: 100%;
-}
-.feature-link{
+.operation-item{
   text-decoration: none;
   color: #043558;
   line-height: 30px;
-  padding: 10px 20px;
+  padding: 5px;
   display: inline-block;
-  width: 100%;
+  width: 80%;
+  border: 1px solid #1155cb;
+  margin-top: 30px;
+  text-align: left;
 }
-.module-feature.active .feature-link,
-.module-subfeature.active .feature-link,
-.feature-link:hover{
+.operation-next{
+  content: ' ';
+  display: none;
+  width: 24px;
+  height: 30px;
+  line-height: 28px;
+  position: absolute;
+  top: 100%;
+  left: 50%;
+  color: #1155cb;
+  margin-left: -12px;
+  z-index: 1;
+  font-size: 28px;
+  text-align: center;
+}
+.operation-feature:not(:last-child) .operation-item .operation-next{
+  display: inline-block;
+}
+.operation-next-add .add-action{
+  display: none;
+  font-size: 10px;
+  border-radius: 50%;
+  width: 16px;
+  height: 16px;
+  line-height: 13px;
+  background: #ffffff;
+  margin-top: 0;
+  margin-bottom: 0;
+  margin-left: -8px;
+  left: 50%;
+  position: absolute;
+  top: 3px;
+  z-index: 1;
+}
+.operation-next:hover .operation-next-add .add-action{
+  display: inline-block;
+}
+.operation-feature.active .operation-item,
+.operation-subfeature.active .operation-item,
+.operation-item:hover{
   background: #f5f5f5;
 }
-.module-sublist .feature-link{
+.operation-sublist .operation-item{
   padding-left: 40px;
-}
-.module-sublist{
-  padding: 0;
 }
 .feature-icon{
   font-size: 1.3em;
@@ -110,8 +170,5 @@ export default {
   width: 20px;
   text-align: center;
   display: inline-block;
-}
-.feature-expand{
-  float: right;
 }
 </style>