+++ CHANGELOG.md
@@ -8,6 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.3.0 - 2022-10-27
+
+### Changed
+* Designer view list styles and actions (add/remove/insert between)
+* Minor style fixes across components
+
+### Added
+* Operation list filter by name/tag/version
+
 ## 0.2.0 - 2022-10-26
 
 ### Changed